requirements.txt
openai==1.58.1
pytelegrambotapi==4.25.0
python-dotenv==1.0.1
